+++ docs/changelog.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Add params for search
+ - Add opportunity to pass template and params by addr and idtf
+
+## [v0.1.3]
+### Fixed
+ - add output types path
 
 ## [v0.1.2]
 ### Fixed